+++ prepare_serial_reaction_time.py
@@ -153,7 +153,7 @@ class SRTExperiment:
         ]
         if not trials:
             return 0.0
-        return np.mean([t.rt_ms for t in trials])
+        return float(np.mean([t.rt_ms for t in trials]))
 
     def get_learning_effect(self) -> float:
         """Sequential - Random RT (should be ~50-100ms faster for sequential)."""
